Major update to cv2df(). Removed df attributes, built-into df instead.
Be advised: This change will break all old reports that used the cv2df() function.
Experimental information (potential limits, scanspeeds, etc.) are now added to the
return dataframe as columns instead of attributes.
With this information built-in to the dataframe, analysis is simplified using melt
and cast and other grouping functions.